requirements.txt
Flask==0.12
gunicorn==19.7.0
Flask-SQLAlchemy==2.2
flask-script==2.0.6
paho-mqtt==1.3.0
